--- internal/inspect/versions/tools.go
@@ -1,38 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package versions
-
-import "github.com/hashicorp/go-version"
-
-// Between returns if the target contains a version between the lowerBound and
-// higherBound.
-func Between(target, lowerBound, higherBound string) bool {
-	v := MustParse(target)
-
-	return v.GreaterThanOrEqual(MustParse(lowerBound)) &&
-		v.LessThanOrEqual(MustParse(higherBound))
-}
-
-// MustParse must be used with versions that are embedded in the code or generated by us,
-// so we are sure that the format is correct. For versions that are provided by a user
-// we should handle properly the error.
-func MustParse(v string) *version.Version {
-	p, err := version.NewVersion(v)
-	if err != nil {
-		// if we reach this point, there must be a bug so we panic
-		panic(err)
-	}
-	return p
-}
